refactor(core): Move tag tests to `__test__` subdir (no-changelog)
