+++ packages/cli/tsconfig.build.json
@@ -6,5 +6,5 @@
 		"tsBuildInfoFile": "dist/build.tsbuildinfo"
 	},
 	"include": ["src/**/*.ts"],
-	"exclude": ["test/**"]
+	"exclude": ["test/**", "src/**/__test__/**"]
 }
